+++ nodalia-cards.manifest.js
@@ -1,10 +1,10 @@
 export default {
-  "pkgVersion": "1.1.0-alpha.13",
-  "contentSha256_12": "37ffb8af4070",
+  "pkgVersion": "1.1.0-alpha.16",
+  "contentSha256_12": "aeaed27a57c8",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
-  "hacsFile": "nodalia-cards-1.1.0-alpha.13.js"
+  "hacsFile": "nodalia-cards-1.1.0-alpha.16.js"
 };
-export const pkgVersion = "1.1.0-alpha.13";
-export const contentSha256_12 = "37ffb8af4070";
+export const pkgVersion = "1.1.0-alpha.16";
+export const contentSha256_12 = "aeaed27a57c8";
 export const file = "nodalia-cards.bundle.js";